refactor: carry counts as int end-to-end instead of narrowing to uint32
Every list total and status count originates as a Go int — ent's Count() or a
len() — and was narrowed to uint32/uint16 at each layer boundary purely
because the OpenAPI spec declared x-go-type there. That cost a conversion at
~25 sites, each needing a gosec (G115) suppression, several of which had
silently stopped working: golines had wrapped the expressions and pushed the
trailing //nolint off the reported line.

Counts now stay int through db wrappers, the movie/tvshow/bulkimport
services, and the generated response types (the JSON wire format is
unchanged). The conversions that remain are the ones with a real reason: ent
columns deliberately kept narrow (import-scan counters, per-show file_count)
convert at a single choke point each, and platform- or protocol-bounded
values (statfs block counts, a bound listen port, TMDB runtime minutes) carry
a standalone-line //nolint stating the bound — the placement golines cannot
break. The season-pack parser needs neither: parsing its two-digit capture
with ParseUint(…, 16) lets gosec prove the range itself.

+++ internal/media/movie/service_test.go
@@ -220,7 +220,7 @@ var _ = Describe("MovieService unit", Label("unit", "movies"), func() {
 
 			items, total, err := svc.List(ctx, 3, 10)
 			Expect(err).NotTo(HaveOccurred())
-			Expect(total).To(Equal(uint32(42)))
+			Expect(total).To(Equal(42))
 			Expect(items).To(Equal(rows))
 		})
 	})
@@ -253,7 +253,7 @@ var _ = Describe("MovieService unit", Label("unit", "movies"), func() {
 			})
 			Expect(err).NotTo(HaveOccurred())
 			Expect(items).To(HaveLen(1))
-			Expect(total).To(Equal(uint32(5)))
+			Expect(total).To(Equal(5))
 		})
 
 		It("wraps filter errors", func() {
